move _set_mpl_backend to nilearn._utils.helpers
